# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anthropic_stc_backend_internal_models.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 32
        minLength: 2
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    - permissions
    type: object
...
  github_com_anthropic_stc_backend_internal_models.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      user_count:
        type: integer
    type: object
  github_com_anthropic_stc_backend_internal_models.SettingGroupResponse:
    properties:
      group_name:
        type: string
      settings:
        items:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.SettingItemResponse'
        type: array
    type: object
  github_com_anthropic_stc_backend_internal_models.SettingItemResponse:
    properties:
      default_value:
        type: string
      description:
        type: string
      display_name:
        type: string
      is_secret:
        type: boolean
      key:
        type: string
      value:
        type: string
      value_type:
        type: string
    type: object
...
  github_com_anthropic_stc_backend_internal_models.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 32
        minLength: 2
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
  github_com_anthropic_stc_backend_internal_models.UpdateSettingsRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
...
  github_com_anthropic_stc_backend_internal_queue.QueueStats:
    properties:
      completed:
        type: integer
      failed:
        type: integer
      pending:
        type: integer
      running:
        type: integer
      workers:
        type: integer
    type: object
...
  github_com_anthropic_stc_backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Prowl Range Backend API
  title: Prowl Range API
  version: 0.1.0
paths:
...
  /queue/stats:
    get:
      description: 返回任务队列统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_queue.QueueStats'
              type: object
      security:
      - BearerAuth: []
      summary: 队列统计
      tags:
      - queue
  /queue/tasks:
    get:
      description: 返回最近的任务列表
      parameters:
      - default: 20
        description: 数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 最近任务列表
      tags:
      - queue
    post:
      description: 创建一个 demo 任务（sleep 5s 模拟工作）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 入队示例任务
      tags:
      - queue
  /roles:
    get:
      description: 获取全部角色
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.RoleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: 创建角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - roles
  /roles/{id}:
    delete:
      description: 删除指定角色
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - roles
    get:
      description: 根据 ID 获取角色详情
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.RoleResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - roles
  /settings/system:
    get:
      description: 按分组返回所有系统配置项，敏感字段值会被遮盖
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.SettingGroupResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取所有系统配置
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: 批量更新配置项，敏感字段传入遮盖值时会跳过不更新
      parameters:
      - description: 配置更新
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 批量更新系统配置
      tags:
      - settings
  /settings/system/{group}:
    get:
      description: 获取指定分组的系统配置项
      parameters:
      - description: 分组名称
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_anthropic_stc_backend_internal_models.SettingItemResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_anthropic_stc_backend_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取分组配置
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"